// src/types/settings.ts
export interface Settings {
  full_name: string;
  access_token: string;
  expiration_date: string;
  email: string;
  game_version: number;
  game_version_user: number;
  download_path_ck1: string;
  download_path_ck2: string;
  last_selected_game: boolean;
  free_game_version_user: number;
  free_game_version: number;
  membership: string;
} 